+++ test/features/test_gradient_estimation.cpp
@@ -65,7 +65,7 @@ TEST (PCL, IntensityGradientEstimation)
       cloud_xyzi.points.push_back (p);
     }
   }
-  cloud_xyzi.width = static_cast<std::uint32_t> (cloud_xyzi.size ());
+  cloud_xyzi.width = cloud_xyzi.size ();
   PointCloud<PointXYZI>::ConstPtr cloud_ptr = cloud_xyzi.makeShared ();
 
   // Estimate surface normals
